MOD: load ed25519 crypto in domains package
