+++ crypto/crypto_ed25519.go
@@ -1,0 +1,9 @@
+// +build ed25519 !custom_alg
+
+package crypto
+
+const KeySize = 32
+
+type Key [KeySize]byte
+type Response [32]byte
+type Commitment Key
