--- rpc/ed25519_test.go
@@ -1,9 +1,0 @@
-// +build ed25519 !custom_alg
-
-package rpc
-
-import "github.com/MixinNetwork/mixin/crypto/ed25519"
-
-func init() {
-	ed25519.Load()
-}
